+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## [1.0.1] - 2017.09.25
-- arjan/decorator lib updated
+- arjan/decorator lib has been updated
 
 ## [1.0.0] - 2017.08.30